requirements.txt
appdirs==1.4.4
asgiref==3.5.2
astroid==2.9.0
attrs==21.4.0
autograd==1.4
beautifulsoup4==4.11.1
bleach==5.0.1
boto3==1.26.22
botocore==1.29.22
bs4==0.0.1
bulma==1.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
cssselect==1.1.0
cycler==0.11.0
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.4
Django==4.1.3
django-admin==2.0.1
django-admin-black==1.0.0
django-admin-volt==1.0.4
django-allauth==0.46.0
django-bona-blog==1.1.4
django-ckeditor==6.2.0
django-cors-headers==3.13.0
django-countries==7.2.1
django-crispy-forms==1.14.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==21.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-rest-auth==0.9.5
django-six==1.0.4
django-storages==1.13.1
django-summernote==0.8.20.0
django-svelte==0.1.6
django-taggit==2.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docopt==0.6.2
ecommercetools==0.42.3
env==0.1.0
fake-useragent==0.1.11
filelock==3.4.2
fonttools==4.37.0
future==0.18.2
gapandas==1.0
google-api-core==2.8.2
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
huggingface-hub==0.9.0
idna==3.4
importlib-metadata==4.12.0
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
kiwisolver==1.4.4
lazy-object-proxy==1.6.0
Lifetimes==0.11.3
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mccabe==0.6.1
mysqlclient==2.1.1
numpy==1.23.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==21.3
pandas==1.4.3
parse==1.19.0
patsy==0.5.2
Pillow==9.3.0
pipenv==2022.1.8
pipreqs==0.4.11
platformdirs==2.4.0
pluggy==1.0.0
protobuf==4.21.5
psycopg2==2.9.5
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycausalimpact==0.1.1
pycodestyle==2.8.0
pycparser==2.21
pyee==8.2.2
PyJWT==2.6.0
pylint==2.12.1
pyparsing==3.0.7
pyppeteer==1.0.2
pyquery==1.4.3
pytest==7.1.1
python-dateutil==2.8.2
python-decouple==3.1
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
requests-html==0.10.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.0
screen==1.0.1
six==1.16.0
sklearn==0.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
statsmodels==0.13.2
stripe==4.2.0
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
torch==1.12.1
tqdm==4.64.0
transformers==4.21.1
typed-ast==1.5.0
typing_extensions==4.3.0
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.13.0
virtualenv-clone==0.5.7
w3lib==2.0.1
webencodings==0.5.1
websockets==10.3
wrapt==1.11.2
xlwt==1.3.0
yarg==0.1.9
zipp==3.8.1
